requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
schedule==1.2.0
groq==0.31.1